Improve UX of password reset
As reported on:

https://github.com/thoughtbot/clearance/issues/250

By default, the text on the password reset page read:

> We will email you a link to reset your password.

This implies that the system is already in the process of sending an
email,
and that no further action is required by the user. We have had multiple
users of our application fail to complete the password reset flow
because of
this confusion.

Here's how these users got stuck:

* Visit the login page.
* Enter their email address and an incorrect password.
* The application denies login with "Bad email or password" error.
* User gives up and clicks "Forgot password?" link.
* Reset password page loads with the text "We will email you a link to
 reset your password.". User thinks they are done. (After all, they
 already provided their email address on the login page.)
* User never completes the reset password form. Telephones support
 complaining their password is not being reset.

Our fix was to change the password reset text to:

> To be emailed a link to reset your password, please enter your email
> address.