Remove sign up link from sign in flash error
There is a configuration option that allows for sign up to easily be
disabled. With this enabled, the `sign_up_path` route helper is not
defined. This causes a 500 when a user login fails because the default
I18n message uses `sign_up_path` for interpolation.

I opted to remove the interpolation, and thus the sign up link, from the
default sign in failure message. This is perhaps somewhat less helpful,
but it makes for a working out of the box experience for people who
disable sign up.

An alternative would be to add another flash message specifically for
this case and add conditionals to `DefaultSignInGuard` (or a helper
object) to handle it. I felt that the increased complexity of that
solution wasn't worth it.