Fix build failures on Rails 4.1.rc1
Rails 4.1 now includes jbuilder 2.0+ in its default gemfile. Move the
jbuilder 1.2 dependency from the clearance `gemfile` into the appraisals
for Rails < 4.1 and add jbuilder 2.0 to the Rails 4.1 appraisal. The
sdoc dependency was also moved into the appraisal definitions because
it is a dependency only because it is included in the rails default
gemfile. This leaves the gemfile for development dependencies only.

Renamed the `create_migration` method in the Clearance install generator
because it was [conflicting with a method][1] that was added into the
class hierarchy in Rails 4.1.

[1]: https://github.com/rails/rails/blob/master/railties/lib/rails/generators/migration.rb#L63