Support Rails 4.2
Most of the changes necessary here were test-suite-only issues. The lone
production-impactful change was adding support for
`ActionMailer#deliver_later` in the `PasswordsController`. This will
automatically use the queue configured with Active Job in order to
background the sending of email. With no queue configured, it will be
delivered synchronously. The old `#deliver` method still works, but
generates deprecation warnings.

The rest of the changes were related to the test suite:
* Add a Rails 4.2 appraisal
* Update cucumber steps to remove unnecessary gems from generated apps.
* Simplify appraisal dependencies thanks to the above.
* Fixed `forgeries_controller_spec`. This is still a brittle way to test
  this functionality but at least it works across Rails versions.
* Removed deprecation related to test suite ordering in the test app
  that is loaded.

There remains a single (repeated) deprecation when running specs on 4.2
which comes from RSpec and will be addressed by upgrading to
`rspec-rails` 3.1 (see [`rspec-rails` issue]. That work will be in a
separate pull-request.

[`rspec-rails` issue]: https://github.com/rspec/rspec-rails/issues/1187