Rename `authorize` filter to `require_login`
This name better expresses the intent of the filter and has the
advantage of not conflicting with the `authorize` method provided by
pundit or wading into the sometimes hairy line demarcating authorization
from authentication.

Clearance users should migrate from `:authorize` to `require_login` as
the former will be removed in 2.0. Be sure to catch reference to
`skip_before_filter :authorize` or `skip_before_action :authorize`,
which the deprecation cannot catch.

addresses #503, #436, and #239