Deprecate `Clearance.root`
This appears to have been a convenience method used only in a single
generator. While the replacement is uglier it has the distinct advantage
of not leaking into the public API of the gem.

It's unlikely anyone is using this method externally, but I added a
deprecation just in case. The method will be removed in 2.0.