Replace Cucumber with RSpec acceptance specs
The replacement spec does much the same as the features were doing
before, but in a (hopefully) more maintainable manner. We generate a
fresh rails app, install clearance, and run clearance's generated specs.

I kept the acceptance spec in its own suite because the output of it is
a bit odd and I wanted it separate from our usual tests. It will output
the results of the generated specs followed by the result of the single
acceptance spec.