Prevent redirect loops when still using `authorize`
The `authorize` before filter is deprecated in favor of `require_login`.
However, just switching Clearance's internal filters is not sufficient.
We're dependent on users updating any `before_filter :authorize` calls.

If they still have `before_filter :authorize` in their application
controller, then they will see a deprecation on calls to authorize but
the method will still run (it's aliased to `require_login`). This may
cause them to be redirected to sign in. Sign in is set to
`skip_before_filter :require_login`, but will happily still run
authorize if instructed to by the application controller. Then you're
stuck in a redirect loop.

By duplicating the `skip_before_filter` calls to also skip `authorize`
we can be sure this doesn't happen.