README Refresh
I took a pass through the README file and edited for brevity,
organization, and correctness. I eliminated the various lists of
overrideable methods as these were out of date and not particularly
helpful in themselves. We already point the users to the overridable
classes. Over time I'd like to move the overriding and extending
documentation into the yarddoc itself or into a website that has recipes
for various behaviors people often request. For now, I've slimmed some
of it down but left most of it.

I also removed reference to the `deny_access` matcher as it will soon be
removed (to another gem) and I don't suggest its use.