Allow User model to be reloaded in development
When a `user_model` is configured in a Clearance initializer, a
reference to that class is immediately saved off. If that class is
changed, Clearance will not know to automatically reload the class as
Rails does automatically for classes in development.

This change introduces a `to_prepare` block to the Engine that is
responsible for forcing the configured user class to be reloaded.
`to_prepare` runs once per request in development and only at startup in
other environments.