Fix mailer delivery expectation
3ca7a6b20a73f90b68fbfdb5a794d6caa1d941ef introduced a change in
`visitor_resets_password_spec.rb.tt` to break up the expectation on the
email's body into an `html_part` and `text_part`. The block that finds
the email however was only considering the match to the `text_part`
because it was not chained to the other conditions, so the subject and
html part were effectively ignored.

This commit chains all the conditions together to ensure they are
finding the right thing.