Test against Rails 5 master
Master includes a fix for https://github.com/rails/rails/issues/23645,
which is not yet included in a beta release.

When updating to master, I also had to explicitly set the ActiveJob
queue adapter to `inline` in tests, as it now defaults to `async`.

With these changes, the non-acceptance test suite is now green and has
no deprecations. Next up - the acceptance tests.