Allow Clearance::BackDoor to take a block
If `User#to_param` is overriden the `Clearance::BackDoor` will fail
because it only looks the user up by id. By passing a block, this
functionality can be replaced to be more flexible. For instance, with a
User model that looks like:

    class User < ActiveRecord::Base
      def to_param
        username
      end
    end

You can specific the following backdoor in
`config/environments/test.rb`:

    config.middleware.use Clearance::BackDoor do |username|
      Clearance.configuration.user_model.find_by(username: username)
    end