Use mime-types 2 for versions tested with Ruby 1.9
Older versions of Rails will happily allow mime-types 3.0+ to be used.
That gem has a dependency on mime-types-data, which requires Ruby 2.0 or
newer.

The additions to the appraisal file here will keep mime-types pinned to
a version compatible with 1.9.3 for these versions of Rails. An
alternative would be to add additional appraisals for each of these
rails versions that only applied to 1.9.3, leaving no restriction on the
mime-types gem for the existing appraisals.

This seemed like more overhead than necessary as we don't directly
interact with the mime-types gem.