Use correct extension for erb migration templates
The rails generator processes these as erb no matter what they are named - which
is why this was working previously despite the confusing `.rb` extension.

This change makes it more clear to developers and their text editors that they
are indeed looking at code which is intended to be erb-processed before it is
copied into it's final resting place.